requirements.txt
discord.py==1.6.0
aiohttp==3.7.4.post0
sentry-sdk==1.0.0
aioredis==1.3.1
